requirements.txt
flask
line-bot-sdk
pymongo